requirements.txt
annoy==1.17.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
Deprecated==1.2.18
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
future==1.0.0
graphtools==1.5.3
Jinja2==3.1.6
joblib==1.5.1
keopscore==2.3
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
packaging==25.0
pacmap==0.8.0
pandas==2.3.1
phate==1.0.11
pillow==11.3.0
pybind11==3.0.0
PyGSP==0.5.1
pykeops==2.3
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
s_gd2==1.8.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
tasklogger==1.2.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
trimap==1.1.4
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
wheel==0.45.1
wrapt==1.17.2
